fix: fix Kotlin compile errors in VMDisplayScreen + VNCClient
VMDisplayScreen.kt: canvas.drawImageRect requires paint= param and
dstSize must be IntSize (unit), not geometry.Size.

VNCClient.kt: ByteArray slots can't hold Int > 127; 255 must be
255.toByte() (= 0xFF as signed byte) for VNC pixel format message.

+++ app/src/main/java/com/ryzix/vm/vnc/VNCClient.kt
@@ -150,11 +150,11 @@ class VNCClient {
         msg[6] = 0  // big-endian = false
         msg[7] = 1  // true-colour = true
         // Red: max=255, shift=16
-        msg[8] = 0; msg[9] = 255
+        msg[8] = 0; msg[9] = 255.toByte()
         // Green: max=255, shift=8
-        msg[10] = 0; msg[11] = 255
+        msg[10] = 0; msg[11] = 255.toByte()
         // Blue: max=255, shift=0
-        msg[12] = 0; msg[13] = 255
+        msg[12] = 0; msg[13] = 255.toByte()
         msg[14] = 16 // red-shift
         msg[15] = 8  // green-shift
         msg[16] = 0  // blue-shift
